Added function_exists( 'compiler' ) check for function compiler()
I wrote code that hooks into WP-SCSS from the child theme and causes both the parent and child theme to compile. I was getting a "cannot redeclare function compiler()" error.

I wrapped function compiler() with a function_exists( 'compiler' ) check and that solved the problem. Would you mind putting this in the next release?

Love the plugin by the way <3.